Hide implementation of cancelling intent
