+++ demo-extended/src/main/java/com/novoda/downloadmanager/demo/extended/DemoNotificationCustomiser.java
@@ -7,7 +7,6 @@ import android.support.v4.app.NotificationCompat;
 
 import com.novoda.downloadmanager.demo.R;
 import com.novoda.downloadmanager.lib.DownloadManager;
-import com.novoda.downloadmanager.lib.DownloadReceiver;
 import com.novoda.downloadmanager.lib.NotificationCustomiser;
 
 class DemoNotificationCustomiser implements NotificationCustomiser {
@@ -24,9 +23,9 @@ class DemoNotificationCustomiser implements NotificationCustomiser {
         PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
         NotificationCompat.Action action = new NotificationCompat.Action(R.drawable.ic_notif_info, "View", pendingIntent);
         builder.addAction(action);
-        Intent cancelIntent = new Intent(DownloadManager.ACTION_CANCEL, null, context, DownloadReceiver.class);
-        cancelIntent.putExtra(DownloadManager.EXTRA_BATCH_ID, batchId);
+        Intent cancelIntent = DownloadManager.createCancelBatchIntent(batchId, context);
         PendingIntent pendingCancelIntent = PendingIntent.getBroadcast(context, 0, cancelIntent, PendingIntent.FLAG_UPDATE_CURRENT);
         builder.addAction(com.novoda.downloadmanager.R.drawable.dl__ic_action_cancel, context.getString(com.novoda.downloadmanager.R.string.dl__cancel), pendingCancelIntent);
     }
+
 }
